+++ src/Mail/Domain/Handler/MailHandler.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Mail\Domain\Handler;
 
 use App\Mail\Domain\Message\MailMessage;
+use Symfony\Bridge\Twig\Mime\TemplatedEmail;
 use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
@@ -24,11 +25,21 @@ class MailHandler
      */
     public function __invoke(MailMessage $mailMessage): void
     {
-        $email = (new Email())
-            ->from($mailMessage->getSender())
-            ->to($mailMessage->getRecipient())
-            ->subject($mailMessage->getTitle())
-            ->text($mailMessage->getContent());
+        if($mailMessage->isTemplated()) {
+            $email = (new TemplatedEmail())
+                ->subject($mailMessage->getTitle())
+                ->from($mailMessage->getSender())
+                ->to($mailMessage->getRecipient())
+                ->htmlTemplate($mailMessage->getContent())
+                ->context($mailMessage->getContext());
+        }
+        else {
+            $email = (new Email())
+                ->from($mailMessage->getSender())
+                ->to($mailMessage->getRecipient())
+                ->subject($mailMessage->getTitle())
+                ->text($mailMessage->getContent());
+        }
 
         $this->mailer->send($email);
     }